add autoremove and autoadd functionality to newrole management
    ALTER TABLE newrole ADD autoremove DEFAULT FALSE,
                        ADD autoadd DEFAULT TRUE;

+++ cogs/management.py
@@ -6,6 +6,7 @@ import discord
 from utils.utils import pluralize, integer
 from utils.checks import need_db
 from cogs.base import BaseCog
+import config
 
 
 def date(argument):
@@ -24,7 +25,7 @@ class Management(BaseCog):
     @commands.group(invoke_without_command=True)
     @commands.guild_only()
     @commands.has_permissions(manage_roles=True)
-    async def newrole(self, ctx, role: discord.Role = None):
+    async def newrole(self, ctx, *, role: discord.Role = None):
         """Automatically give new members a role.
 
         Call without a role to see current role.
@@ -58,8 +59,59 @@ class Management(BaseCog):
     @newrole.command()
     @commands.guild_only()
     @commands.has_permissions(manage_roles=True)
+    async def autoadd(self, ctx, toggle: bool = None):
+        """Check or set whether to automatically add the new role. Default: enabled
+
+        If enabled, the new role will automatically be added if a user has all other roles removed.
+        """
+        state = await ctx.con.fetchval('''
+            SELECT autoadd FROM newrole WHERE guild_id = $1
+            ''', ctx.guild.id)
+        if state is None:
+            await ctx.send('New role is not set for this guild.')
+        elif toggle is None:
+            await ctx.send(f'New role autoremoval is set to {"enabled" if state else "disabled"}')
+        else:
+            async with ctx.con.transaction():
+                await ctx.con.execute('''
+                    UPDATE newrole SET autoadd = $1 WHERE guild_id = $2
+                    ''', toggle, ctx.guild.id)
+            await ctx.send(f'New role removal is now {"enabled" if toggle else "disabled"}')
+
+    @need_db
+    @newrole.command()
+    @commands.guild_only()
+    @commands.has_permissions(manage_roles=True)
+    async def autoremove(self, ctx, toggle: bool = None):
+        """Check or set whether to automatically remove the new role. Default: disabled
+
+        If enabled, the new role will automatically be removed when the user is assigned another role.
+        """
+        state = await ctx.con.fetchval('''
+            SELECT autoremove FROM newrole WHERE guild_id = $1
+            ''', ctx.guild.id)
+        if state is None:
+            await ctx.send('New role is not set for this guild.')
+        elif toggle is None:
+            await ctx.send(f'New role autoremoval is set to {"enabled" if state else "disabled"}')
+        else:
+            async with ctx.con.transaction():
+                await ctx.con.execute('''
+                    UPDATE newrole SET autoremove = $1 WHERE guild_id = $2
+                    ''', toggle, ctx.guild.id)
+            await ctx.send(f'New role removal is now {"enabled" if toggle else "disabled"}')
+
+    @need_db
+    @newrole.command()
+    @commands.guild_only()
+    @commands.has_permissions(manage_roles=True)
     async def off(self, ctx):
         """Stop automically giving new members a role."""
+        exists = await ctx.con.fetchval('''
+            SELECT EXISTS(*) FROM newrole WHERE guild_id = $1
+            ''', ctx.guild.id)
+        if not exists:
+            return await ctx.send('New role is not set for this guild.')
         async with ctx.con.transaction():
             await ctx.con.execute('''
                 DELETE FROM newrole WHERE guild_id = $1
@@ -153,6 +205,27 @@ class Management(BaseCog):
                         DELETE FROM newrole WHERE guild_id = $1
                         ''', member.guild.id)
         await member.add_roles(role, reason='New Member')
+
+    async def on_member_update(self, before, after):
+        """Remove new role when user is assigned another role."""
+        len_before = len(before.roles)
+        len_after = len(after.roles)
+        if not after.guild.me.guild_permissions.manage_roles or \
+                len(before.roles) == len(after.roles):
+            return
+        async with self.bot.db_pool.acquire() as con:
+            settings = await con.fetchrow('''
+                SELECT * FROM newrole WHERE guild_id = $1
+                ''', after.guild.id)
+        if settings is None or (not settings['autoremove'] and not settings['autoadd']):
+            return
+        newrole = discord.utils.get(before.roles, id=settings['role_id'])
+        if settings['autoremove'] and len_after > len_before:
+            if newrole not in after.roles:
+                return
+            await after.remove_roles(newrole, reason=f'{config.prefix}newrole autoremove')
+        elif settings['autoadd'] and len_after == 1:  # all roles removed, only has @everyone
+            await after.add_roles(newrole, reason=f'{config.prefix}newrole autoadd')
 
 
 def setup(bot):
